+++ internal/archive/tar_builder_test.go
@@ -26,6 +26,7 @@ func TestTARBuilder_Build(t *testing.T) {
 		description         string
 		archive             *archive.TARBuilder
 		lastArchiveInfo     func(backupPaths []string) archive.Info
+		ignorePatterns      []*regexp.Regexp
 		backupPaths         []string
 		expected            func(filename string) error
 		expectedArchiveInfo func(backupPaths []string) archive.Info
@@ -58,6 +59,9 @@ func TestTARBuilder_Build(t *testing.T) {
 					},
 				}
 			},
+			ignorePatterns: []*regexp.Regexp{
+				regexp.MustCompile(`^.*\~\$.*$`),
+			},
 			backupPaths: func() []string {
 				d, err := ioutil.TempDir("", "toglacier-test")
 				if err != nil {
@@ -69,6 +73,10 @@ func TestTARBuilder_Build(t *testing.T) {
 				}
 
 				if err := ioutil.WriteFile(path.Join(d, "file2"), []byte("file2 test"), os.ModePerm); err != nil {
+					t.Fatalf("error creating temporary file. details %s", err)
+				}
+
+				if err := ioutil.WriteFile(path.Join(d, "~$file3"), []byte("file3 test"), os.ModePerm); err != nil {
 					t.Fatalf("error creating temporary file. details %s", err)
 				}
 
@@ -583,7 +591,7 @@ func TestTARBuilder_Build(t *testing.T) {
 				lastArchiveInfo = scenario.lastArchiveInfo(backupPaths)
 			}
 
-			filename, archiveInfo, err := scenario.archive.Build(lastArchiveInfo, backupPaths...)
+			filename, archiveInfo, err := scenario.archive.Build(lastArchiveInfo, scenario.ignorePatterns, backupPaths...)
 			if scenario.expectedError == nil && scenario.expected != nil {
 				if err = scenario.expected(filename); err != nil {
 					t.Errorf("unexpected archive content (%s). details: %s", filename, err)